Change links containing master to main
NOTES:

In 07.BLAST_tutorial.Rmd which is located in the educe website repo, there is a link to workshops_other/blob/master/... <-- didn't change

There is a zip file called master.zip inside workshop_data, that is linked to in the 07.BLAST_tutorial.Rmd <-- didn't change and didn't move this file to educer

Similarly there's a lot of fasta files in workshop_data that are used a lot in the 07.BLAST_tutorial.Rmd file. I didn't move them to educer, but I did change the link to main.

In 09.Git_tutorial.Rmd and 11.GitHub_tutorial.Rmd changed every instance of master to main (not parts of links). But they have pictures with the word master in them.